functional reversion

+++ Arcade.RPG/Configs/Models/Viewport.cs
@@ -9,8 +9,8 @@ public class Zoom {
 
 public class Viewport {
     public Entity? Subject { get; set; } = null;
-    public float TileBaseWidth { get; set; } = 32;
-    public float TileBaseHeight { get; set; } = 32;
+    public int TileBaseWidth{ get; set; } = 32;
+    public int TileBaseHeight { get; set; } = 32;
     public float TileX { get; set; } = 0;
     public float TileY { get; set; } = 0;
     public float TileXRadius { get; set; } = 15;
